moe_final_v0.6.0_sept24 based fused_moe_kernel

+++ byte_infer_perf/llm_perf/backends/ROCM/fused_moe/csrc/moe_ops.h
@@ -6,8 +6,9 @@ void topk_softmax(torch::Tensor& topk_weights, torch::Tensor& topk_indices,
                   torch::Tensor& token_expert_indices,
                   torch::Tensor& gating_output);
 
-void moe_align_block_size(torch::Tensor topk_ids, int num_experts,
-                          int block_size, torch::Tensor sorted_token_ids,
+void moe_align_block_size(torch::Tensor topk_ids, int64_t num_experts,
+                          int64_t block_size, torch::Tensor sorted_token_ids,
                           torch::Tensor experts_ids,
                           torch::Tensor num_tokens_post_pad);
+
 void silu_and_mul(torch::Tensor &out, torch::Tensor &input);
